Allow arbitrary coinage systems
The default coinage system remains USD. To enable other currency
systems, add a definition to currencies.json and pass the --currency
option with the currency's ISO code.

The verbiage in the code is Not Great. I don't know if there's a general
term for dollars and cents. Also, if a currency's small units aren't a
hundredth of the main denomination, this will produce incorrect results!