Use a divisor field to Currency
This is the divisor used by the currency's fractional unit. There are
only a handful of money systems out there that don't use 100, but they
do exist.

We need to know this as the coin values are in the fractional units, but
the program is required to accept and output the currency's primary
unit! This is not ideal because we should not be doing floating point
math on currency values here. Our vending machine is (hopefully) not
calculating interest rates :^)